refactor: rename dynamic and static entity packages to adhere to naming conventions
