+++ src/main/scala/io/github/srs/model/entity/dynamicentity/Actuator.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model.entity.dynamic_entity
+package io.github.srs.model.entity.dynamicentity
 
 trait Actuator[E <: DynamicEntity]:
   def act(entity: E): E
