+++ src/main/scala/io/github/srs/model/entity/staticentity/StaticEntity.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model.entity.static_entity
+package io.github.srs.model.entity.staticentity
 
 import io.github.srs.model.entity.*
 import io.github.srs.model.validation.Validation
