+++ src/test/scala/io/github/srs/model/entity/staticentity/StaticEntityTest.scala
@@ -1,6 +1,6 @@
-package io.github.srs.model.entity.static_entity
+package io.github.srs.model.entity.staticentity
 
-import io.github.srs.model.entity.static_entity.StaticEntity
+import io.github.srs.model.entity.staticentity.StaticEntity
 import io.github.srs.model.entity.{ Orientation, Point2D }
 import io.github.srs.model.validation.DomainError
 import org.scalatest.EitherValues.*
